add pagination, security, admin API

+++ NoticeBoard/src/main/java/atom/id/noticeboard/services/MessageService.java
@@ -11,10 +11,14 @@ import atom.id.noticeboard.exceptions.NotFoundException;
 import atom.id.noticeboard.exceptions.TopicHasNoMessagesException;
 import atom.id.noticeboard.repositories.MessageRepository;
 import atom.id.noticeboard.repositories.TopicRepository;
+import atom.id.noticeboard.security.MyUserDetails;
 import jakarta.validation.constraints.NotNull;
 import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
 import lombok.experimental.FieldDefaults;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -31,6 +35,7 @@ public class MessageService {
     TopicRepository topicRepository;
     MappingTopicUtils mappingTopicUtils;
 
+
     public TopicWithMessagesDto saveNewMessage(MessageDto dto, String topicId) {
         if(!dto.getAuthor().isEmpty() && !dto.getText().isEmpty() && dto.getCreated() != null) {
             Topic topic = topicRepository.findById(topicId).orElseThrow(() -> new NotFoundException("Invalid topic ID"));
@@ -45,7 +50,18 @@ public class MessageService {
     }
 
     public TopicWithMessagesDto updateMessage(MessageDto dto, String topicId) {
+        String getAuthUsername = MyUserDetails.getAuthUsername();
+        if(!dto.getAuthor().equals(getAuthUsername))
+            throw new InvalidInputException("This is not your message");
+        return getUpdatedTopicWIthMessages(dto, topicId);
+    }
 
+    public TopicWithMessagesDto updateAnyMessage(MessageDto dto, String topicId) {
+
+        return getUpdatedTopicWIthMessages(dto, topicId);
+    }
+
+    private TopicWithMessagesDto getUpdatedTopicWIthMessages(MessageDto dto, String topicId) {
         if(!dto.getAuthor().isEmpty() && !dto.getText().isEmpty() && dto.getCreated() != null) {
             Topic topic = topicRepository.findById(topicId).orElseThrow(() ->
                     new NotFoundException("Invalid topic ID"));
@@ -62,14 +78,35 @@ public class MessageService {
     }
 
     public void deleteMessage(@NotNull String id) {
+        Authentication auth = SecurityContextHolder.getContext().getAuthentication();
+
         Message message = messageRepository.findById(id).orElseThrow();
+        if(!auth.getName().equals(message.getAuthor()))
+            throw new InvalidInputException("This is not your message");
+
         if(topicRepository.findById(message.getTopic().getId()).get().getMessages().size() <= 1)
             throw new TopicHasNoMessagesException("Topic should not have less than 1 messages");
 
         messageRepository.deleteById(id);
     }
 
+    public List<Message> getMessagesByTopicId(String topicId, PageRequest pageRequest) {
+
+        return Optional.of(messageRepository.findAllByTopicId(topicId,pageRequest)).orElseGet(ArrayList::new);
+    }
+
     public List<Message> getMessagesByTopicId(String topicId) {
+
         return Optional.of(messageRepository.findAllByTopicId(topicId)).orElseGet(ArrayList::new);
+    }
+
+    public void deleteAnyMessage(@NotNull String id) {
+        Message message = messageRepository.findById(id).orElseThrow(() ->
+                new InvalidInputException("message with this id was not found"));
+
+        if(topicRepository.findById(message.getTopic().getId()).get().getMessages().size() <= 1)
+            topicRepository.deleteById(message.getTopic().getId());
+
+        messageRepository.deleteById(id);
     }
 }
